+++ vidsjs/modules/index.js
@@ -6,3 +6,5 @@ exports.items = require('./items');
 exports.utils = require('./utils');
 exports.view = require('./view');
 exports.seen = require('./seen');
+exports.users = require('./users');
+exports.loginchecker = require('./loginchecker');